+++ dtex_utility.h
@@ -26,10 +26,10 @@ void dtex_relocate_spr(struct ej_package*, struct int_array* array, int tex_idx,
 
 void dtex_relocate_c2_key(struct dtex_c2*, struct ej_package*, struct int_array* array, struct dtex_img_pos* src, struct dtex_img_pos* dst);
 
-void dtex_relocate_pic_part(int16_t part_src[8], struct dtex_inv_size* src_sz, struct dtex_rect* src_rect, 
+void dtex_relocate_pic_part(float part_src[8], struct dtex_inv_size* src_sz, struct dtex_rect* src_rect, 
 	struct dtex_inv_size* dst_sz, struct dtex_rect* dst_rect, int rotate, float trans_vb[16], float dst_vb[8]);
 
-void dtex_get_pic_src_rect(int16_t* src, struct dtex_rect* rect);
+void dtex_get_pic_src_rect(float* src, struct dtex_rect* rect);
 
 #endif // dynamic_texture_utility_h
